Changed sources into a file called sources.txt.
Added getSources() that returns list of source urls.

Changed name of xmlToJson() to goXmlToJson to better represent that
it is turning a custom Go object into a json byte array.

Changed the ServeHTTP() function to handle a list of sources.
It now serves a list of json objects, where each sources RSS feed is
in "data":